KeyLogger/requirements.txt
keyboard==0.13.5
python-dotenv==0.20.0
